+++ SandboxiePlus/MiscHelpers/Common/SortFilterProxyModel.h
@@ -14,14 +14,8 @@ public:
 	{
 		m_bHighLight = false;
 		m_iColumn = 0;
-		m_pView = NULL;
 
 		this->setSortCaseSensitivity(Qt::CaseInsensitive);
-	}
-
-	void setView(QTreeView* pView)
-	{
-		m_pView = pView;
 	}
 
 	bool filterAcceptsRow(int source_row, const QModelIndex & source_parent) const
@@ -86,124 +80,16 @@ public slots:
 		QRegularExpression RegExp(ExpStr, (iOptions & CFinder::eCaseSens) != 0 ? QRegularExpression::NoPatternOption : QRegularExpression::CaseInsensitiveOption);
 
 		QModelIndex idx;
-		//if (m_pView) idx = m_pView->currentIndex();
 		m_iColumn = Col;
 		m_bHighLight = (iOptions & CFinder::eHighLight) != 0;
 		setFilterKeyColumn(Col); 
 		setFilterRegularExpression(RegExp);
-		//if (m_pView) m_pView->setCurrentIndex(idx);
 		if (m_bHighLight)
 			emit layoutChanged();
-	}
-
-	void SelectNext()
-	{
-		if (!m_pView)
-			return;
-
-		bool next = true;
-		QModelIndex idx = m_pView->currentIndex();
-		if (!(next = idx.isValid()))
-			idx = index(0, 0);
-
-		//if (QApplication::keyboardModifiers() & Qt::ControlModifier)
-		if (QApplication::keyboardModifiers() & Qt::ShiftModifier)
-			idx = FindPrev(idx, next);
-		else
-			idx = FindNext(idx, next);
-
-		if (idx.isValid())
-			m_pView->setCurrentIndex(idx);
-		else
-			QApplication::beep();
 	}
 
 protected:
 	bool		m_bHighLight;
 	int			m_iColumn;
-	QTreeView*	m_pView;
-
-	bool		MatchCell(QModelIndex idx, int column)
-	{
-		QModelIndex tmp = idx.sibling(idx.row(), column);
-
-		QString str = data(tmp, filterRole()).toString();
-		if (str.contains(filterRegularExpression()))
-			return true;
-		return false;
-	}
-
-	bool		MatchRow(QModelIndex idx)
-	{
-		if (m_iColumn != -1)
-			return MatchCell(idx, m_iColumn);
-
-		for(int col = 0; col < columnCount(idx); col++) {
-			if (MatchCell(idx, col))
-				return true;
-		}
-		return false;
-	}
-
-	QModelIndex	FindNext(QModelIndex idx, bool next = false)
-	{
-		if (MatchRow(idx) && !next)
-			return idx;
-
-		if (hasChildren(idx))
-		{
-			int numRows = rowCount(idx);
-			for (int count = 0; count < numRows; count++) {
-				QModelIndex tmp = FindNext(index(count, 0, idx));
-				if (tmp.isValid())
-					return tmp;
-			}
-		}
-
-		do {
-			QModelIndex par = parent(idx);
-
-			int numRows = rowCount(par);
-			for (int count = idx.row() + 1; count < numRows; count++) {
-				QModelIndex tmp = FindNext(index(count, 0, par));
-				if (tmp.isValid())
-					return tmp;
-			}
-
-			idx = par;
-		} while (idx.isValid());
-
-		return QModelIndex();
-	}
-
-	QModelIndex	FindPrev(QModelIndex idx, bool next = false)
-	{
-		if (MatchRow(idx) && !next)
-			return idx;
-
-		if (hasChildren(idx))
-		{
-			int numRows = rowCount(idx);
-			for (int count = numRows-1; count >= 0; count++) {
-				QModelIndex tmp = FindNext(index(count, 0, idx));
-				if (tmp.isValid())
-					return tmp;
-			}
-		}
-
-		do {
-			QModelIndex par = parent(idx);
-
-			int numRows = rowCount(par);
-			for (int count = idx.row() - 1; count >= 0; count--) {
-				QModelIndex tmp = FindNext(index(count, 0, par));
-				if (tmp.isValid())
-					return tmp;
-			}
-
-			idx = par;
-		} while (idx.isValid());
-
-		return QModelIndex();
-	}
 };
+